feat: unique handler per endpoint
