+++ src/handler/get-voting-power.ts
@@ -3,22 +3,21 @@ import {
     APIGatewayProxyResult
 } from 'aws-lambda/trigger/api-gateway-proxy'
 import {Context} from 'aws-lambda'
-import {log} from '../config/logging'
-import {Database} from './db/database'
-import {throwError} from './error'
+import {log} from '../../config/logging'
+import {Database} from '../db/database'
+import {throwError} from '../error'
 
 /**
  * Shape of the AWS Gateway event the handler has available
  * https://github.com/awsdocs/aws-lambda-developer-guide/blob/main/sample-apps/nodejs-apig/event.json
  */
-
 export const handler = async (
     event: APIGatewayProxyEvent,
     context: Context
     // eslint-disable-next-line @typescript-eslint/require-await
 ): Promise<APIGatewayProxyResult> => {
     log.info(
-        'handler id: %s, name: %s',
+        'getVotingPower id: %s, name: %s',
         context.awsRequestId,
         context.functionName
     )
